+++ src/DigiSign.php
@@ -28,7 +28,7 @@ use Psr\SimpleCache\CacheInterface;
 
 final class DigiSign implements EndpointInterface
 {
-    public const VERSION = '1.8.0';
+    public const VERSION = '1.9.0';
     public const API_BASE = 'https://api.digisign.org';
     public const API_BASE_TESTING = 'https://api.digisign.digital.cz';