+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+### Added
+- Add `EnvelopeRecipientTemplate.bankIdScopes`
+
 ## [1.7.0]
 ### Added
 - Add `AccountCertificateEndpoint` and `Certificate` resource